fix: make sure `all_goals` restores state on failure
This PR modifies `all_goals` so that in recovery mode it commits changes to the state only for those goals for which the tactic succeeds (while preserving the new message log state). Before, we were trusting that failing tactics left things in a reasonable state, but now we roll back and admit the goal. This fixes a bug where we were rolling back only the metacontext state and not the tactic state, leading to an inconsistent state (a goal list with metavariables not in the metacontext).

Alternatively we could stop on the first error, but while writing a tactic script it is helpful to see what the tactic did to each goal. There is some non-monotonicity here however since tactics can solve for metavariables that appear in successive goals, and conceivably a later goal succeeds only if a previous one does. Given that the non-monotonicity is limited to recovery mode (which is for example the RHS and not the LHS of the `<;>` combinator), we think this is acceptible.

Closes #7883