feat: in app elaborator, beta reduce when substituting arguments into type
This PR modifies the app elaborator so that as arguments are elaborated, they are substituted into the remaining function type with beta reduction. This is useful for eliminator-like applications. In an eliminator, if a user provides an explicit motive, it was the case that the expression would be beta reduced, but now it is beta reduced like in the `@[elab_as_elim]` procedure:
```lean4
example (xs : List α) : xs ++ [] = xs :=
  List.recOn xs
    (motive := fun xs' => xs' ++ [] = xs')
    (nil := rfl)
    (cons := fun x xs' ih => by rw [List.cons_append, ih])
/-           ^
Before:
⊢ ∀ (head : α) (tail : List α),
  (fun xs' ↦ xs' ++ [] = xs') tail → (fun xs' ↦ xs' ++ [] = xs') (head :: tail)
After (and before when `motive` is deleted):
⊢ ∀ (head : α) (tail : List α),
  tail ++ [] = tail → head :: tail ++ [] = head :: tail
-/
```