feat: use omission dots for hidden let values in Infoview
This PR changes the behavior of `pp.showLetValues` to use a hoverable `⋯` to hide let values. This is now false by default, and there is a new option `pp.showLetValues.threshold` for allowing small expressions to be shown anyway. For tactic metavariables, `pp.showLetValues.tactic.threshold` is also considered, and this is by default set to the maximal value, since in tactic states local values are usually significant.