feat: characterize when T-division equals zero
This PR adds the lemma `Int.tdiv_eq_zero_iff_natAbs_lt_or_eq_zero`
which shows that T-division equals zero iff the absolute value
of the numerator is less than the denominator, or the denominator equals
zero.

This was used in reasoning about the MSB of `BitVec.sdiv`. The strategy was
superced, but this lemma was extracted from the earlier proof.