doc: also mention the full syntax for the deprecated attribute
The other day, I wanted to apply the full form (i.e., with both a custom
message and a deprecation date) in mathlib and failed to guess the correct
order of these two. The attribute doc-string seemed to suggest that combining
them was impossible (which was luckily false).

To prevent this from re-occurring, clarify the doc-string by
- mentioning that a custom message and deprecation date can be combined
- what the expected order of these two is