feat: Meta.letToHave
This PR adds a module for transforming `let`s to `have`s when we can detect that they are not dependent. The transformation is applied to all declarations defined using `def`/`theorem`/`instance`/`opaque`.

The compiler does not seem to recognize join points that use `have` instead of `let`, so there's a hack where `let`s named `__do_jp` are preserved. We'll revisit this when the new compiler is in place.